[FIX/#43] 홈 페이지 빌드 오류 수정
